Make changes to allow tests to run in project, security token tests all passing
