+++ src/contract_wrappers/tokens/__tests__/erc20_wrapper.test.ts
@@ -2,9 +2,9 @@
 import { BigNumber } from '@0x/utils';
 import { mock, instance, reset, when, verify, objectContaining } from 'ts-mockito';
 import { Web3Wrapper } from '@0x/web3-wrapper';
-import { DetailedERC20Contract } from '@polymathnetwork/abi-wrappers';
+import { ERC20DetailedContract } from '@polymathnetwork/abi-wrappers';
 import ContractWrapper from '../../contract_wrapper';
-import DetailedERC20Wrapper from '../detailed_erc20_wrapper';
+import DetailedERC20Wrapper from '../erc20_detailed_wrapper';
 import { MockedCallMethod, MockedSendMethod, getMockedPolyResponse } from '../../../test_utils/mocked_methods';
 import { valueToWei, weiToValue } from '../../../utils/convert';
 
@@ -12,11 +12,11 @@ describe('ERC20TokenWrapper', () => {
   // Declare ERC20TokenWrapper object
   let target: DetailedERC20Wrapper;
   let mockedWrapper: Web3Wrapper;
-  let mockedContract: DetailedERC20Contract;
+  let mockedContract: ERC20DetailedContract;
 
   beforeAll(() => {
     mockedWrapper = mock(Web3Wrapper);
-    mockedContract = mock(DetailedERC20Contract);
+    mockedContract = mock(ERC20DetailedContract);
 
     const myContractPromise = Promise.resolve(instance(mockedContract));
     target = new DetailedERC20Wrapper(instance(mockedWrapper), myContractPromise);
